Centralize VoxelAABB definition
