+++ common/src/node.rs
@@ -4,6 +4,7 @@ use std::ops::{Index, IndexMut};
 
 use serde::{Deserialize, Serialize};
 
+use crate::collision_math::Ray;
 use crate::dodeca::Vertex;
 use crate::graph::{Graph, NodeId};
 use crate::lru_slab::SlotId;
@@ -202,4 +203,210 @@ fn populate_node(graph: &mut Graph, node: NodeId) {
             .unwrap_or_else(NodeState::root),
         chunks: Chunks::default(),
     });
+}
+
+/// Represents a discretized region in the voxel grid contained by an axis-aligned bounding box.
+pub struct VoxelAABB {
+    // The bounds are of the form [[x_min, x_max], [y_min, y_max], [z_min, z_max]], using voxel coordinates with a one-block
+    // wide margins added on both sides. This helps make sure that that we can detect if the AABB intersects the chunk's boundaries.
+    bounds: [[u8; 2]; 3],
+}
+
+impl VoxelAABB {
+    /// Returns a bounding box that is guaranteed to cover a given radius around a ray segment. Returns None if the
+    /// bounding box lies entirely outside the chunk.
+    pub fn from_ray_segment_and_radius(
+        layout: &ChunkLayout,
+        ray: &Ray,
+        tanh_distance: f32,
+        radius: f32,
+    ) -> Option<VoxelAABB> {
+        // Convert the ray to grid coordinates
+        let grid_start =
+            na::Point3::from_homogeneous(ray.position).unwrap() * layout.dual_to_grid_factor();
+        let grid_end = na::Point3::from_homogeneous(ray.ray_point(tanh_distance)).unwrap()
+            * layout.dual_to_grid_factor();
+        // Convert the radius to grid coordinates using a crude conservative estimate
+        let max_grid_radius = radius * layout.dual_to_grid_factor();
+        let mut bounds = [[0; 2]; 3];
+        for axis in 0..3 {
+            let grid_min = grid_start[axis].min(grid_end[axis]) - max_grid_radius;
+            let grid_max = grid_start[axis].max(grid_end[axis]) + max_grid_radius;
+            let voxel_min = (grid_min + 1.0).floor().max(0.0);
+            let voxel_max = (grid_max + 1.0)
+                .floor()
+                .min(layout.dimension() as f32 + 1.0);
+
+            // When voxel_min is greater than dimension or voxel_max is less than 1, the cube does not intersect
+            // the chunk.
+            if voxel_min > layout.dimension() as f32 || voxel_max < 1.0 {
+                return None;
+            }
+
+            // We convert to u8 here instead of earlier because out-of-range voxel coordinates can violate casting assumptions.
+            bounds[axis] = [voxel_min.floor() as u8, voxel_max.floor() as u8];
+        }
+
+        Some(VoxelAABB { bounds })
+    }
+
+    /// Iterator over grid points contained in the region, represented as ordered triples
+    pub fn grid_points(
+        &self,
+        axis0: usize,
+        axis1: usize,
+        axis2: usize,
+    ) -> impl Iterator<Item = (u8, u8, u8)> {
+        let bounds = self.bounds;
+        (bounds[axis0][0]..bounds[axis0][1]).flat_map(move |i| {
+            (bounds[axis1][0]..bounds[axis1][1])
+                .flat_map(move |j| (bounds[axis2][0]..bounds[axis2][1]).map(move |k| (i, j, k)))
+        })
+    }
+
+    /// Iterator over grid lines intersecting the region, represented as ordered pairs determining the line's two fixed coordinates
+    pub fn grid_lines(&self, axis0: usize, axis1: usize) -> impl Iterator<Item = (u8, u8)> {
+        let bounds = self.bounds;
+        (bounds[axis0][0]..bounds[axis0][1])
+            .flat_map(move |i| (bounds[axis1][0]..bounds[axis1][1]).map(move |j| (i, j)))
+    }
+
+    /// Iterator over grid planes intersecting the region, represented as integers determining the plane's fixed coordinate
+    pub fn grid_planes(&self, axis: usize) -> impl Iterator<Item = u8> {
+        self.bounds[axis][0]..self.bounds[axis][1]
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::collections::HashSet;
+
+    use crate::math;
+
+    use super::*;
+
+    /// Any voxel AABB should at least cover a capsule-shaped region consisting of all points
+    /// `radius` units away from the ray's line segment. This region consists of two spheres
+    /// and a cylinder. We only test planes because covered lines and points are a strict subset.
+    #[test]
+    fn voxel_aabb_coverage() {
+        let dimension = 12;
+        let layout = ChunkLayout::new(dimension);
+
+        // Pick an arbitrary ray by transforming the positive-x-axis ray.
+        let ray = na::Rotation3::from_axis_angle(&na::Vector3::z_axis(), 0.4).to_homogeneous()
+            * math::translate_along(&na::Vector3::new(0.2, 0.3, 0.1))
+            * &Ray::new(na::Vector4::w(), na::Vector4::x());
+
+        let tanh_distance = 0.2;
+        let radius = 0.1;
+
+        // We want to test that the whole capsule-shaped region around the ray segment is covered by
+        // the AABB. However, the math to test for this is complicated, so we instead check a bunch of
+        // spheres along this ray segment.
+        let num_ray_test_points = 20;
+        let ray_test_points: Vec<_> = (0..num_ray_test_points)
+            .map(|i| {
+                math::lorentz_normalize(
+                    &ray.ray_point(tanh_distance * (i as f32 / (num_ray_test_points - 1) as f32)),
+                )
+            })
+            .collect();
+
+        let aabb =
+            VoxelAABB::from_ray_segment_and_radius(&layout, &ray, tanh_distance, radius).unwrap();
+
+        // For variable names and further comments, we use a tuv coordinate system, which
+        // is a permuted xyz coordinate system.
+
+        // Test planes in all 3 axes.
+        for t_axis in 0..3 {
+            let covered_planes: HashSet<_> = aabb.grid_planes(t_axis).collect();
+
+            // Check that all uv-aligned planes that should be covered are covered
+            for t in 0..=dimension {
+                if covered_planes.contains(&t) {
+                    continue;
+                }
+
+                let mut plane_normal = na::Vector4::zeros();
+                plane_normal[t_axis] = 1.0;
+                plane_normal[3] = layout.grid_to_dual(t);
+                let plane_normal = math::lorentz_normalize(&plane_normal);
+
+                for test_point in &ray_test_points {
+                    assert!(
+                        math::mip(test_point, &plane_normal).abs() > radius.sinh(),
+                        "Plane not covered: t_axis={t_axis}, t={t}, test_point={test_point:?}",
+                    );
+                }
+            }
+        }
+
+        // Test lines in all 3 axes
+        for t_axis in 0..3 {
+            let u_axis = (t_axis + 1) % 3;
+            let v_axis = (u_axis + 1) % 3;
+            let covered_lines: HashSet<_> = aabb.grid_lines(u_axis, v_axis).collect();
+
+            // For a given axis, all lines have the same direction, so set up the appropriate vector
+            // in advance.
+            let mut line_direction = na::Vector4::zeros();
+            line_direction[t_axis] = 1.0;
+            let line_direction = line_direction;
+
+            // Check that all t-aligned lines that should be covered are covered
+            for u in 0..=dimension {
+                for v in 0..=dimension {
+                    if covered_lines.contains(&(u, v)) {
+                        continue;
+                    }
+
+                    let mut line_position = na::Vector4::zeros();
+                    line_position[u_axis] = layout.grid_to_dual(u);
+                    line_position[v_axis] = layout.grid_to_dual(v);
+                    line_position[3] = 1.0;
+                    let line_position = math::lorentz_normalize(&line_position);
+
+                    for test_point in &ray_test_points {
+                        assert!(
+                            (math::mip(test_point, &line_position).powi(2)
+                                - math::mip(test_point, &line_direction).powi(2))
+                            .sqrt()
+                                > radius.cosh(),
+                            "Line not covered: t_axis={t_axis}, u={u}, v={v}, test_point={test_point:?}",
+                        );
+                    }
+                }
+            }
+        }
+
+        // Test points
+        let covered_points: HashSet<_> = aabb.grid_points(0, 1, 2).collect();
+
+        // Check that all points that should be covered are covered
+        for x in 0..=dimension {
+            for y in 0..=dimension {
+                for z in 0..=dimension {
+                    if covered_points.contains(&(x, y, z)) {
+                        continue;
+                    }
+
+                    let point_position = math::lorentz_normalize(&na::Vector4::new(
+                        layout.grid_to_dual(x),
+                        layout.grid_to_dual(y),
+                        layout.grid_to_dual(z),
+                        1.0,
+                    ));
+
+                    for test_point in &ray_test_points {
+                        assert!(
+                            -math::mip(test_point, &point_position) > radius.cosh(),
+                            "Point not covered: x={x}, y={y}, z={z}, test_point={test_point:?}",
+                        );
+                    }
+                }
+            }
+        }
+    }
 }
